Extracts PDF failure status badge into a separate component
Moves the failure tooltip logic, state management, and positioning code from PdfActivityCard into a new PdfFailureStatusBadge component to improve maintainability and simplify the parent component.
